# api/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Comprehensive Electronic Medical Record system for hospitals",
        "title": "Hospital EMR System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@hospital-emr.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "http://www.hospital-emr.com/license"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/encounters": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a paginated list of encounters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "encounters"
                ],
                "summary": "List encounters",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by patient ID",
                        "name": "patient_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by provider ID",
                        "name": "provider_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new clinical encounter for a patient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "encounters"
                ],
                "summary": "Create a new encounter",
                "parameters": [
                    {
                        "description": "Encounter information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_encounter.CreateEncounterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Encounter"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_common_errors.AppError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_common_errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/v1/encounters/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed information about an encounter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "encounters"
                ],
                "summary": "Get encounter by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Encounter ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Encounter"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_common_errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/v1/encounters/{id}/complete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mark an encounter as completed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "encounters"
                ],
                "summary": "Complete encounter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Encounter ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Encounter"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_common_errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/v1/encounters/{id}/diagnoses": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a diagnosis to an encounter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "encounters"
                ],
                "summary": "Add diagnosis",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Encounter ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Diagnosis",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_encounter.AddDiagnosisRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Diagnosis"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_common_errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/v1/encounters/{id}/notes": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a clinical note to an encounter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "encounters"
                ],
                "summary": "Add clinical note",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Encounter ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Clinical note",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_encounter.AddClinicalNoteRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.ClinicalNote"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_common_errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/v1/encounters/{id}/status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the status of an encounter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "encounters"
                ],
                "summary": "Update encounter status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Encounter ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Encounter"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_common_errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/v1/encounters/{id}/vitals": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Record vital signs for an encounter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "encounters"
                ],
                "summary": "Record vital signs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Encounter ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Vital signs",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_encounter.RecordVitalSignsRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.VitalSign"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_common_errors.AppError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_hospital-emr_backend_internal_common_errors.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_hospital-emr_backend_internal_models.Diagnosis": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "diagnosed_by": {
                    "type": "string"
                },
                "diagnosis_type": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.DiagnosisType"
                },
                "encounter_id": {
                    "type": "string"
                },
                "icd10_code": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "onset_date": {
                    "type": "string"
                },
                "resolved_date": {
                    "type": "string"
                },
                "severity": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
...
        "github_com_hospital-emr_backend_internal_models.Encounter": {
            "type": "object",
            "properties": {
                "admission_date": {
                    "type": "string"
                },
                "chief_complaint": {
                    "type": "string"
                },
                "clinical_notes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.ClinicalNote"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "department": {
                    "type": "string"
                },
                "diagnoses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Diagnosis"
                    }
                },
                "discharge_date": {
                    "type": "string"
                },
                "encounter_number": {
                    "type": "string"
                },
                "encounter_type": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.EncounterType"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "orders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Order"
                    }
                },
                "patient": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Patient"
                },
                "patient_id": {
                    "type": "string"
                },
                "priority": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Priority"
                },
                "procedures": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Procedure"
                    }
                },
                "provider": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.User"
                },
                "provider_id": {
                    "type": "string"
                },
                "reason_for_visit": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.EncounterStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                },
                "vital_signs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.VitalSign"
                    }
                }
            }
        },
...
        "github_com_hospital-emr_backend_internal_models.EncounterType": {
            "type": "string",
            "enum": [
                "outpatient",
                "inpatient",
                "emergency",
                "wellness",
                "telehealth"
            ],
            "x-enum-varnames": [
                "EncounterTypeOutpatient",
                "EncounterTypeInpatient",
                "EncounterTypeEmergency",
                "EncounterTypeWellness",
                "EncounterTypeTelehealth"
            ]
        },
...
        "internal_encounter.AddClinicalNoteRequest": {
            "type": "object",
            "required": [
                "note_type"
            ],
            "properties": {
                "assessment": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "note_type": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.NoteType"
                },
                "objective": {
                    "type": "string"
                },
                "plan": {
                    "type": "string"
                },
                "subjective": {
                    "type": "string"
                }
            }
        },
        "internal_encounter.AddDiagnosisRequest": {
            "type": "object",
            "required": [
                "description",
                "diagnosis_type",
                "icd10_code"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "diagnosis_type": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.DiagnosisType"
                },
                "icd10_code": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "onset_date": {
                    "type": "string"
                },
                "severity": {
                    "type": "string"
                }
            }
        },
        "internal_encounter.CreateEncounterRequest": {
            "type": "object",
            "required": [
                "admission_date",
                "encounter_type",
                "patient_id",
                "provider_id"
            ],
            "properties": {
                "admission_date": {
                    "type": "string"
                },
                "chief_complaint": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "encounter_type": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.EncounterType"
                },
                "location": {
                    "type": "string"
                },
                "patient_id": {
                    "type": "string"
                },
                "priority": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Priority"
                },
                "provider_id": {
                    "type": "string"
                },
                "reason_for_visit": {
                    "type": "string"
                }
            }
        },
        "internal_encounter.RecordVitalSignsRequest": {
            "type": "object",
            "properties": {
                "blood_pressure_diastolic": {
                    "type": "integer"
                },
                "blood_pressure_systolic": {
                    "type": "integer"
                },
                "heart_rate": {
                    "type": "integer"
                },
                "height": {
                    "type": "number"
                },
                "notes": {
                    "type": "string"
                },
                "oxygen_saturation": {
                    "type": "number"
                },
                "pain": {
                    "type": "integer"
                },
                "respiratory_rate": {
                    "type": "integer"
                },
                "temperature": {
                    "type": "number"
                },
                "temperature_unit": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}